+++ Changes.md
@@ -7,6 +7,8 @@
 
 ### Changes
 
+* clamd: add check.authenticated, check.private_ip, check.local_ip option 
+
 * dkim_sign: improved log messages #2499
 
 * ehlo_hello_message: config/ehlo_hello_message can be used to overwrite the EHLO/HELO msg replacing `, Haraka is at your service` #2498